jupyterlab_img_gen/scipy/pip_requirements.txt
python-dotenv
alembic
chardet
defusedxml
distlib
idna
jupyterthemes
keplergl
lesscpy
mako
pandas-read-xml
pangres
pillow
ply
pyarrow
pyproj
python-editor
requests
traittypes
urllib3
wikidata2df
xmltodict
zipfile36
pydeck
plotly
streamlit
panel
vaex
flask
voila
python-telegram-bot
lckr-jupyterlab-variableinspector
ipyleaflet
jupyterlab-git==0.30.0b1
jupyterlab_widgets
jupyter-server-proxy
jupyterlab_execute_time
jupyterlab-kite>=2.0.2
jupyterlab-system-monitor

